pyproject.toml
[project]
name = "project-akhir"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
